refactor(server): split handler/messaging.rs
2,395 lines, and the roadmap's standing complaint that it was "still the worst
file, still holding the hot path". Two of the three obvious moves are done; the
third is recorded rather than rushed.

## What moved

**Tests to `messaging/tests.rs`.** 1,674 of the 2,395 lines were the test
module. The production code it covers is ~720, and that is the part worth being
able to read without scrolling past a suite three times its length.

**Pure decision helpers to `messaging/decisions.rs`.** `shape_response_history`,
`json_byte_len`, `second_send_blocked`, `token_aged`, `evict_aged_token`,
`token_still_evictable` and the history cap — free functions with no `self` and
no I/O. The send path now reads as a sequence of decisions rather than decisions
interleaved with the predicates behind them, and each predicate is separately
testable, which matters because their edge cases (an aged token, a still-in-use
context lock) are exactly what the mutation suite probes.

`messaging.rs` (2,395) -> `mod.rs` (644) + `decisions.rs` (101) + `tests.rs`
(1,677).

## What did not

`mod.rs` is still over the 500-line limit, and the reason is a single function:
`send_message_inner` is ~530 lines carrying the entire send path —
interceptors, extension and streaming capability checks, id and metadata
validation, capacity admission, three-phase stale-token eviction under
alternating read/write locks, and task creation with the FIX(#8) ordering that
inserts the cancellation token before the store write.

That is a hot-path refactor with real deadlock and race surface, not a file
move. Doing it badly is worse than not doing it, so both remaining entries are
recorded in `.file-length-baseline` with a note saying exactly what is left and
why — which is what that baseline is for: making a decision visible instead of
silent.

Verified: 1,608 server tests pass at all features (unchanged), clippy
`-D warnings` clean at all features, fmt clean, all six Format-job gates exit 0.
